+++ src/for-tests/utils/introspection/index.ts
@@ -1,4 +1,5 @@
 // created from 'create-ts-index'
 
+export * from './MXYZ_Interfaces';
 export * from './OZ_Interfaces';
 export * from './makeInterfaceId';