+++ CHANGELOG.md
@@ -6,9 +6,11 @@ As of 12/13/2022, this repo has been renamed from "nftc-open-contracts" to "nftc
 
 ## Version -- 1.5.next [Not published]
 - TODO
+
+## Version -- 1.5.21
 - Add ERC20 interfaces to the OZ Interface definition 
 - also write supports interface test for ERC20_165 Mock.
-
+- Add manifold version of shouldSupportsInterface unit test generator.
 
 ## Version -- 1.5.20
 - Implement ERC165 compliant version of ERC20.